requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.13.1
drf-spectacular==0.21.1
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.3.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
zipp==3.6.0
